Move published record view methods into their own file
