+++ core/views/__init__.py
@@ -7,3 +7,4 @@ from .record import *
 from .transformation_scenario import *
 from .datatables import *
 from .field_mapper import *
+from .published import *
